moved formatting into separate module
