+++ app/createHold/index.js
@@ -9,14 +9,13 @@ const app = express.Router( { mergeParams : true } );
 
 app.use( ( req, res ) => {
   logger.info( `${ config.app } initial context based on ${ req.originalUrl }` );
-  const context = { 
-    routePath: req.route.path,
-    logonUrl: config.logonUrl,
-    itemsOutUrls: config.itemsOutUrls,
-    holdsUrls: config.holdsUrls,
+
+  const context = {
     barcodeOrUsername: req.params.user,
+    logonUrl: config.logonUrl,
     password: req.params.pwd,
-    rememberMe: false
+    rememberMe: false,
+    url: config.createHoldUrl
   };
 
   logger.debug( `getting results for msg...` );
@@ -25,21 +24,19 @@ app.use( ( req, res ) => {
     const delim = "--";
     let formattedData = "";
     if ( results.length > 0 ) {
-      
-      logger.debug( `formatting results for msg...` );      
+
+      logger.debug( `formatting results for msg...` );
       results.forEach( item => {
         formattedData += formattedData === "" ? `${ delim }` : `\n${ delim }`;
-        formattedData += ( context.routePath === "/due/:user/:pwd" ) ? 
-                        `${ item.dueDate }${ delim }${ item.id }${ delim }${ item.title }${ delim }${ item.renewal }` :
-                        `${ item.position }${ delim }${ item.id }${ delim }${ item.title }`;
+        formattedData += `${ item.position }${ delim }${ item.id }${ delim }${ item.title }`;
       });
     }
     const messageText = formattedData !== "" ? formattedData : "No Results...";
-    
+
     logger.debug( `sending response...` );
     res.send( messageText );
   })
-  .catch( error => { res.send( error ) } );  
+  .catch( error => { res.send( error ) } );
 
 });
